Move HTTP fetch, pump, and file-staging utilities to gosub_net
Moves fetch.rs (fetch_response_top/complete, ProgressReader), pump.rs
(spawn_pump, PumpCfg, PumpTargets), fs_utils.rs (temp_path_for,
stage_temp_path_for), and the Waiter/stream_to_bytes helpers from
gosub_engine::net into gosub_net::net. Engine files become thin re-exports.
Adds tempfile dep and tokio macros+net features to gosub_net.
