+++ crates/gosub_net/Cargo.toml
@@ -21,9 +21,13 @@ thiserror = "1"
 futures-core = "0.3"
 futures-util = "0.3"
 parking_lot = "0.12"
+tempfile = "3.21.0"
+
+[dev-dependencies]
+cow-utils = "0.1.3"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio = { version = "1.52.3", features = ["rt", "io-util", "sync", "time"] }
+tokio = { version = "1.52.3", features = ["rt", "io-util", "sync", "time", "macros", "net"] }
 tokio-util = { version = "0.7.16", features = ["io"] }
 tokio-stream = { version = "0.1.17", features = ["sync"] }
 reqwest = { version = "0.13.3", features = ["rustls", "stream", "blocking"] }
