+++ crates/gosub_net/src/net/mod.rs
@@ -1,6 +1,9 @@
 pub mod events;
+pub mod fetch;
+pub mod fs_utils;
 pub mod null_emitter;
 pub mod observer;
+pub mod pump;
 pub mod request_ref;
 pub mod shared_body;
 pub mod types;
